+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 
 
-- [SeqKit v0.15.1](https://github.com/shenwei356/seqkit/releases/tag/v0.15.1)
-[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.15.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.15.1)
+- [SeqKit v0.16.0](https://github.com/shenwei356/seqkit/releases/tag/v0.16.0)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.16.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.16.0)
     - `seqkit grep/locate`:
         - fix bug of `-m` when querying contains letters not in alphabet, usually for protein sequences. [#178](https://github.com/shenwei356/seqkit/issues/178), [#179](https://github.com/shenwei356/seqkit/issues/179)
         - onply search on positive strand when searching unlimited or protein sequences
         - new flag `--immediate-output`
+        - **using multi-threads to accelerate searching by sequences while allowing mismatch**.
     - `seqkit locate`:
         - removing debug info for `-r` introduced in a0f6b6e. [#180](https://github.com/shenwei356/seqkit/issues/180)
     - `seqkit fx2tab`:
@@ -15,7 +16,7 @@
     - `seqkit subseq`:
         - fix bug for BED with empty columns. [#195](https://github.com/shenwei356/seqkit/issues/195)
     - `seqkit genautocomplete`: 
-        - support bash|zsh|fish|powershell.
+        - **support bash|zsh|fish|powershell**.
 - [SeqKit v0.15.0](https://github.com/shenwei356/seqkit/releases/tag/v0.15.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.15.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.15.0)
     - `seqkit grep/locate`: update help message.